pyproject.toml
[project]
name = "aslive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "dashscope>=1.20.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.135.3",
    "uvicorn>=0.44.0",
    "python-multipart>=0.0.26",
]
